requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
googlemaps==4.10.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
ollama==0.1.5